Improve the UI. Use scrollable area.
Wrap the menu(s) into a 'frame'. This creates a scrollable area. It means
the user can navigate inside a list that is potentially larger than the
screen size.

This is useful when the user's screen is too small or its directory
contains too many files.

On top of that, make the panel on the right to be flexible.

Demo:
[![asciicast](https://asciinema.org/a/49rEj0FoYvutHR5Uece23O45W.svg)](https://asciinema.org/a/49rEj0FoYvutHR5Uece23O45W)